Move channel & thread syncer to sync utils file
